Remove black bars, commonize formats, hardcode w/h
This effort was to remove black bars to improve design, as they
conflicted with semi-transparent header/footer.

This has created more negative space, and the best way to counter that
is to adjust glade so that the draw areas have constraints apply to
them. For now, there is extra, negative space below the camera streams.
We'd like to put performance numbers there.

Also, fix the distortion in the monodepth demo.