Fix list bug, change drawing schedule design
The sample data was originally a to be stored in a fixed size deque. Add
that back it, it was an accident to remove that.

Change the drawing schedule to queue draws only when the graphing
drawing is done. No sense in having a glib call saturate the call stack
trying to finish draws, I believe that was causing severe call stack
issues under certain circumstances.

Performance is much better now.