Timestamp data for accurate graphing
Now that graphing happens when it can be weaved into gtk scheduler, need
to make sure the x axis is accurate. Implemented this by timestamping
data and dropping stale data.

Still, in the first 30 seconds of initialization, there is some
inaccuracy while the graph ramps up. That is next to be solved.