Automatic USB failure detection v1
Gstreamer class will look for evidence USB stopped working; in that
event, restart the gstreamer pipeline.

In general this works well, but there is some memory allocation
somewhere not being cleaned up. Need to do another pass over this to
improve the UX and prevent memory leak. I think its in GLib or the
pipeline resources.